HandsOn8/requirements.txt
torch
transformers
gradio
peft
sentence-transformers
faiss-cpu
streamlit